Keep quick resume type within source limits
