+++ src/domain/mod.rs
@@ -14,7 +14,9 @@ use serde::{Deserialize, Serialize};
 // Actions domain types (workflows, runs, jobs, steps, filters) extracted to
 // keep this file under the source-file-size limit.
 mod actions;
+mod quick_resume;
 pub use actions::*;
+pub use quick_resume::QuickResume;
 
 /// Stable identifier for a repository.
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
@@ -837,21 +839,6 @@ pub enum AgentStatus {
     Waiting,
     Paused,
     Dead,
-}
-
-/// Typed checkbox value for Code Puppy autosave continuation.
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-pub struct QuickResume(pub bool);
-
-impl QuickResume {
-    #[must_use]
-    pub fn enabled(self) -> bool {
-        self.0
-    }
-
-    pub fn toggle(&mut self) {
-        self.0 = !self.0;
-    }
 }
 
 /// An agent is the primary work unit in Jefe.
